New refactor in finance addition (debit and credit are the same)
