+++ src/main/java/com/rift/runners/finances/manager/entity/Finance.java
@@ -8,36 +8,61 @@ package com.rift.runners.finances.manager.entity;
 import java.io.Serializable;
 import java.util.Objects;
 import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.MappedSuperclass;
+import javax.persistence.Table;
+import javax.xml.bind.annotation.XmlRootElement;
 
 /**
  *
  * @author Guilherme
  */
-@MappedSuperclass
+@Entity
+@Table(name = "finances")
+@XmlRootElement
 public class Finance implements Serializable {
 
     private static final long serialVersionUID = 1L;
-	
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "id_finance")
+    private Long id;
     @Column(length = 50, nullable = false)
     private String name;
     @ManyToOne
-    @JoinColumn(name = "id_subject")
+    @JoinColumn(name = "id_subject", nullable = true)
     private Subject subject;
     private String type;
     @Column(name = "finance_value")
     private Double financeValue;
+    @Column(nullable = true)
+    private Boolean paid;
+    //@Column(name = "payable_until")
+    //private Date payableUntil;
 
     public Finance() {
     }
 
-    public Finance(String name, Subject subject, String type, Double financeValue) {
+    public Finance(Long id, String name, Subject subject, String type, Double financeValue, Boolean paid) {
+        this.id = id;
         this.name = name;
         this.subject = subject;
         this.type = type;
         this.financeValue = financeValue;
+        this.paid = paid;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
     }
 
     public String getName() {
@@ -72,13 +97,23 @@ public class Finance implements Serializable {
         this.financeValue = financeValue;
     }
 
+    public Boolean getPaid() {
+        return paid;
+    }
+
+    public void setPaid(Boolean paid) {
+        this.paid = paid;
+    }
+
     @Override
     public int hashCode() {
-        int hash = 7;
-        hash = 13 * hash + Objects.hashCode(this.name);
-        hash = 13 * hash + Objects.hashCode(this.subject);
-        hash = 13 * hash + Objects.hashCode(this.type);
-        hash = 13 * hash + Objects.hashCode(this.financeValue);
+        int hash = 5;
+        hash = 47 * hash + Objects.hashCode(this.id);
+        hash = 47 * hash + Objects.hashCode(this.name);
+        hash = 47 * hash + Objects.hashCode(this.subject);
+        hash = 47 * hash + Objects.hashCode(this.type);
+        hash = 47 * hash + Objects.hashCode(this.financeValue);
+        hash = 47 * hash + Objects.hashCode(this.paid);
         return hash;
     }
 
@@ -100,12 +135,19 @@ public class Finance implements Serializable {
         if (!Objects.equals(this.type, other.type)) {
             return false;
         }
+        if (!Objects.equals(this.id, other.id)) {
+            return false;
+        }
         if (!Objects.equals(this.subject, other.subject)) {
             return false;
         }
         if (!Objects.equals(this.financeValue, other.financeValue)) {
             return false;
         }
+        if (!Objects.equals(this.paid, other.paid)) {
+            return false;
+        }
         return true;
     }
+
 }
